+++ packages/backend-plugin-api/src/services/definitions/RootLoggerService.ts
@@ -14,15 +14,7 @@
  * limitations under the License.
  */
 
-import { PluginDatabaseManager } from '@backstage/backend-common';
-import { createServiceRef } from '../system/types';
+import { LoggerService } from './LoggerService';
 
 /** @public */
-export type DatabaseService = PluginDatabaseManager;
-
-/**
- * @public
- */
-export const databaseServiceRef = createServiceRef<DatabaseService>({
-  id: 'core.database',
-});
+export interface RootLoggerService extends LoggerService {}