+++ plugins/adr-backend/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @backstage/plugin-adr-backend
 
+## 0.2.5-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.18.0-next.0
+  - @backstage/config@1.0.6-next.0
+  - @backstage/catalog-client@1.3.0-next.1
+  - @backstage/catalog-model@1.1.5-next.1
+  - @backstage/errors@1.1.4
+  - @backstage/integration@1.4.2-next.0
+  - @backstage/plugin-adr-common@0.2.5-next.1
+  - @backstage/plugin-search-common@1.2.1-next.0
+
+## 0.2.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.5-next.0
+  - @backstage/catalog-client@1.3.0-next.0
+  - @backstage/backend-common@0.17.0
+  - @backstage/config@1.0.5
+  - @backstage/errors@1.1.4
+  - @backstage/integration@1.4.1
+  - @backstage/plugin-adr-common@0.2.5-next.0
+  - @backstage/plugin-search-common@1.2.0
+
 ## 0.2.4
 
 ### Patch Changes