+++ plugins/explore-backend/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @backstage/plugin-explore-backend
 
+## 0.0.2-next.1
+
+### Patch Changes
+
+- 9f9d279bd1: Updated `README.md` examples
+- Updated dependencies
+  - @backstage/backend-common@0.18.0-next.0
+  - @backstage/config@1.0.6-next.0
+  - @backstage/plugin-explore-common@0.0.1
+  - @backstage/plugin-search-common@1.2.1-next.0
+
+## 0.0.2-next.0
+
+### Patch Changes
+
+- 358554ff3c: Bump `msw` to `^0.49.0`
+- Updated dependencies
+  - @backstage/backend-common@0.17.0
+  - @backstage/config@1.0.5
+  - @backstage/plugin-explore-common@0.0.1
+  - @backstage/plugin-search-common@1.2.0
+
 ## 0.0.1
 
 ### Patch Changes